@shamatar 2020-06-10
> > Oh, it just hit my mind. Your code is actually 6 pushes + staticcall (or 6 pushes + 6 pops), so staticcall cost that you have proposed should be divided by two, no?
> 
> 
> 
> I don't understand your reasoning there... care to elaborate? 
> 
> I mean, both loops run until OOG, not some set number of loops. 

Oh, what I have meant is that each simple piece of code crunches through the 100M gas over N loop iterations, but those iterations include not only staticcall, but an extra code around it. The conclusion looks to be is that staticcall is equal to 6 POPs in Geth, that should be 12 gas.