@holiman 2020-06-11
Ah, no, ignore the N iterations. That's just the bench framework way to wind up with a stable timing. We simply wants both loops to burn through 100M gas in roughly the same time, regardless of what ops are executed. The calling loop will execute fewer loops than the popping loop, but that's fine.