@shamatar 2020-06-11
If this benchmarks results
```
BenchmarkSimpleLoop/identity-precompile-100M
BenchmarkSimpleLoop/identity-precompile-100M-6         	       2	 674733768 ns/op	244293824 B/op	 3816870 allocs/op
BenchmarkSimpleLoop/loop-100M
BenchmarkSimpleLoop/loop-100M-6                        	       2	 666473308 ns/op	   13508 B/op	      41 allocs/op
```
represents that a function with STATICCALL to precompile that does nothing (modified identity) and a function than does just 6 POPs instead take roughly the same time to execute then STATICCALL = 6 POPs = 12 gas?