@holiman 2020-06-11
I have pushed some changes now, which work their way around the alloc problems. With those in place, it appears that `40` is a pretty solid value: 
```
[user@work runtime]$ go test . -run - -bench BenchmarkSimpleLoop 
goos: linux
goarch: amd64
pkg: github.com/ethereum/go-ethereum/core/vm/runtime
BenchmarkSimpleLoop/identity-precompile-100M-6                 2         664909701 ns/op           16828 B/op         63 allocs/op
BenchmarkSimpleLoop/loop-100M-6                                2         676632829 ns/op           14044 B/op         42 allocs/op
BenchmarkSimpleLoop/identity-precompile-10M-6                 18          66192889 ns/op           16006 B/op         61 allocs/op
BenchmarkSimpleLoop/loop-10M-6                                16          68658524 ns/op           13235 B/op         40 allocs/op

```