@shamatar 2020-06-11
Didn't dig deeper, but for similar tests in OpenEthereum

```
crunch through identity for 100M gas                                                                           
                        time:   [151.31 ms 153.17 ms 154.75 ms]
crunch through identity for 100M gas with 40 gas per staticcall                                                                          
                        time:   [1.5607 s 1.6060 s 1.6629 s]
crunch through dummy loop with pops instead of staticcall for 100M gas                        
                        time:   [1.8245 s 1.8524 s 1.8747 s]
```

Now I need to understand how to interpret them, but roughly it looks like it may be a little below 40 gas for OpenEthereum