@holiman 2020-06-12
Cool! Could you link to your code?If you haven't modified identity, I think your measuring a cost of 56 gas instead of 40. So give 26 a spin too.