@holiman 2020-06-12
Also, might want to give it a warm-up run before benchmarking. To load state cache and instantiate everything. My benchmark is the pure runloop without the extra time for initializing state etc.