@shamatar 2020-06-12
The code is [here](https://github.com/matter-labs/openethereum/commit/77471a1d08a0f088dfd3b30802036b3e0fbb38a6). I didn't test other values cause you can optimistically get 40 gas per STATICCALL already implemented if you use "berlin" schedule.