@tkstanczak 2020-06-12
@holiman I think that @shamatar was right here
your benchmark would give different results if you executed it for 1 billion gas for example

I think the better approach would be to set a counter to X and loop that many times and then compare the timings - this would give you the actual gas cost of the static call

you would need to set the gast cost of the static call to the number to make these two benchmarks equal and then you would have to subtract 12 (cost of 6 pops) from this value