@tkstanczak 2020-06-12
by the way - the number that I am getting is something like 350 gas for static call now but I need to dig a bit deeper