@tkstanczak 2020-06-12
looking at this allocs on our side to bring it down - it seems that our code invokes empty account zero transfer touch logic on precompile which slows it down (and a few other things)