@holiman 2020-06-12
I actually did every benchmark with both 100M and 10M, to verify that the results are correlated. I can run with 1Bn gas too, I don't expect any difference. I don't understand what's wrong with my reasoning, could you elaborate what you think is off?