@shamatar 2020-06-12
Do avoid any doubt about what the benchmark measures can we just place a code that runs through the loop for 1M times, and measure how long it would take to run? Irrespective of the gas consumed. Then we can place POPs to make running time equal and price STATICCALL this way.