@shamatar 2020-06-12
The benchmarks as it’s written right now will either stop if it does N iterations with staticcall or M iterations with POPs. 

If prices are the proper then N*cost = M*cost_1 = time * GAS/second, so invariants should be preserved and you can calculate costs if you will read N or M somehow.