@holiman 2020-06-12
I don't think that's a good approach - then you'll price staticcall using pop (and push or dup) as the golden standard, and only those. My approach smears it out a bit, so we compare against the set of all instructions in the loop, some pushes, dup's,  a jump + jumpdest. 