@shamatar 2020-06-12
Dummy cycle can be make out from many opcodes, but if most popular opcodes (as it seems to me) in EVM are not well priced then it’s kind of a separate large problem