@holiman 2020-06-12
Here's with `10 billion`, `100 million` and `10 million` gas. I'd say it' pretty consistent:
```
BenchmarkSimpleLoop/staticcall-identity-10B
BenchmarkSimpleLoop/staticcall-identity-10B-6         	       1	89446314582 ns/op	   17752 B/op	      66 allocs/op
BenchmarkSimpleLoop/loop-1B
BenchmarkSimpleLoop/loop-1B-6                        	       1	72261810479 ns/op	   14968 B/op	      45 allocs/op
BenchmarkSimpleLoop/staticcall-identity-100M
BenchmarkSimpleLoop/staticcall-identity-100M-6       	       2	 704888916 ns/op	   16828 B/op	      63 allocs/op
BenchmarkSimpleLoop/loop-100M
BenchmarkSimpleLoop/loop-100M-6                      	       2	 966903886 ns/op	   13508 B/op	      41 allocs/op
BenchmarkSimpleLoop/staticcall-identity-10M
BenchmarkSimpleLoop/staticcall-identity-10M-6        	      16	  69748344 ns/op	   16019 B/op	      61 allocs/op
BenchmarkSimpleLoop/loop-10M
BenchmarkSimpleLoop/loop-10M-6                       	      15	  73673448 ns/op	   13314 B/op	      40 allocs/op
```