@shamatar 2020-06-12
Inversion between 10B and 100M is interesting