@tkstanczak 2020-06-12
so in Nethermind we are at around 150 gas with the current code

we can go to 41 gas with some significant changes (chosing a different path for precompile calls)
this also assumes some logic that avoids the path for account Touch handling

@holiman are you reporting the account touch event in your benchmark?