@tkstanczak 2020-06-12
also, I think we picked a test where no memory operations are executed - because of the loop with STATICCALL passing 0 length memory access and zero length result write
These operations are not coming for free so maybe worth to also verify a scenario where the result copying is happening (unless this is covered by the precompile base cost?