@shamatar 2020-06-12
How does EVM handle memory extensions and on what point?